Fix podcast keep out-dated issue
For example if user has subscribed an podcast, and haven't upgrade
yet. Then if user search for the podcast, one will get the latest
info. However, when one goes back to check the subscribed podcast,
the info is not latest.

Since we will always add the podcast to DB if not exists yet when user
check podcast search result. We can step further by upgrading an existed
podcast, while keeping the `isSubscribed` field the same as before.